cwd_cam_web_app/requirements.txt
asgiref==3.8.1
certifi==2023.7.22
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
django-environ==0.11.2
django_csp==3.8
djangorestframework==3.15.2
idna==3.4
openlocationcode==1.0.1
pillow==10.4.0
pluscodes==2022.1.3
psycopg==3.2.1
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.0.7
whitenoise==6.7.0
